refactor: rename index.tsx to PascalCase.tsx, add barrel index.tsx for S2-1b components
Matches the Tasklist convention established in S2-1a. Also inlines the
deleted constants (var(--cds-spacing-09), z-index 7999) directly in
CollapsablePanel/styled.ts.
