+++ webapp/client/apps/orchestration-cluster-webapp/src/operate/shared/FiltersPanel/index.tsx
@@ -6,56 +6,4 @@
  * except in compliance with the Camunda License 1.0.
  */
 
-import React, {useEffect, useState} from 'react';
-import {getStateLocally, storeStateLocally} from '#/shared/browser-storage/local-storage';
-import {CollapsablePanel} from '../CollapsablePanel';
-import {Button} from '@carbon/react';
-import {Footer} from './styled';
-
-type Props = {
-	localStorageKey: 'isFiltersCollapsed' | 'isDecisionsFiltersCollapsed' | 'isAuditLogsFiltersCollapsed';
-	children: React.ReactNode;
-	onResetClick?: () => void;
-	isResetButtonDisabled: boolean;
-};
-
-const FiltersPanel: React.FC<Props> = ({children, localStorageKey, onResetClick, isResetButtonDisabled}) => {
-	const isCollapsed = (getStateLocally('operate.panelStates')?.[localStorageKey] as boolean | undefined) ?? false;
-	const [panelState, setPanelState] = useState<'expanded' | 'collapsed'>(isCollapsed ? 'collapsed' : 'expanded');
-
-	useEffect(() => {
-		storeStateLocally('operate.panelStates', {
-			...(getStateLocally('operate.panelStates') ?? {}),
-			[localStorageKey]: panelState === 'collapsed',
-		});
-	}, [panelState, localStorageKey]);
-
-	return (
-		<CollapsablePanel
-			label="Filter"
-			panelPosition="LEFT"
-			maxWidth={320}
-			isCollapsed={panelState === 'collapsed'}
-			onToggle={() => {
-				setPanelState((panelState) => {
-					if (panelState === 'collapsed') {
-						return 'expanded';
-					}
-
-					return 'collapsed';
-				});
-			}}
-			footer={
-				<Footer>
-					<Button kind="ghost" size="sm" disabled={isResetButtonDisabled} type="reset" onClick={onResetClick}>
-						Reset filters
-					</Button>
-				</Footer>
-			}
-		>
-			{children}
-		</CollapsablePanel>
-	);
-};
-
-export {FiltersPanel};
+export {FiltersPanel} from './FiltersPanel';
